whatsapp-image-sender-backend/requirements.txt
flask
flask-cors
joblib
pillow
numpy
scikit-learn
tensorflow
